goesaodc_createRasterStack()

Create a RasterStack from GOES AOD data files for the date and hour specified by datetime. Each RasterLayer contains data from one Advanced Baseline Imager (ABI) scan during the specified time period.

If data for the specified time period is not found in the directory specified by setSatelliteDataDir(), it will be downloaded in order to create the RasterStack.

The Z axis of the RasterStack is a character vector where each element is the time stamp of the scan and has the format YYYYMMDDHHMMSS. This can be accessed using the raster::getZ() function. Names of the RasterStack are also time stamps of the scan, of the format XHH.MM.SS.

The dqfLevel parameter can take a value of:

0 -- High quality retrieval flag

1 -- Medium quality retrieval flag

2 -- Low quality retrieval flag

3 -- No retrieval quality flag

The bbox parameter can be a vector of floats in c(lonLo, lonHi, latLo, latHi) order or the return value from sp::bbox() or raster::extent().

goesaodc_createRasterStack(
  satID = NULL,
  datetime = NULL,
  endTime = NULL,
  var = "AOD",
  res = 0.1,
  bbox = c(-125, -65, 24, 50),
  dqfLevel = NULL,
  timezone = "UTC",
  isJulian = FALSE,
  fileList = NULL,
  verbose = FALSE
)

Arguments

satID
ID of the source GOES satellite (G16 or G17).

datetime
Desired datetime in any Ymd H [MS] format or POSIXct.

endTime
Desired ending time in any Ymd H [MS] format or POSIXct

var
GOES data variable ("AOD, "DQF" or "ID"). Default: "AOD"

res
resolution of raster in degrees. Default: 0.1

bbox
Bounding box for raster, Default: c(-125, -65, 24, 50) CONUS

dqfLevel
Data quality flag level.

timezone
timezone for datetime and optionally endTime.

isJulian
Logical specifying if datetime (and optionally endTime) are Julian formatted

fileList
optional list of files to stack. Useful when working with custom time ranges.

verbose
Logical specifies whether to print stacking information.

Value

RasterStack

Examples

if (FALSE) {
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

satID <- "G17"
datetime <- "2019-10-27 10"
endTime <- "2019-10-27 11"
bbox <- c(-124, -120, 36, 39) # Kincade fire region
dqfLevel <- 3
timezone <- "America/Los_Angeles"

rasterStack <- goesaodc_createRasterStack(
satID = satID,
datetime = datetime,
endTime = endTime,
bbox = bbox,
dqfLevel = dqfLevel,
timezone = timezone,
res = 0.075,
verbose = TRUE)

rasterVis::levelplot(rasterStack, par.settings=BuRdTheme())
 }