Download GOES-16 or GOES-17 AOD data

Download all GOES 16 or 17 NetCDF files for the given datetime to the directory specified by setSatelliteDataDir(). If datetime is specified to the hour and fullDay is not explicitly set to TRUE', only files for that hour will be downloaded. If the optional endTime is specified, all files that exist for the time period between datetime and endTime will be downloaded. If timezone is not specified, "UTC" is assumed.

The vector of files returned includes all files in satelliteDataDir that match the requested date whether they were download previously or with this call.

goesaodc_downloadAOD(satID = NULL, datetime = NULL, endTime = NULL,
  timezone = "UTC", isJulian = FALSE, verbose = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/")

Arguments

satID
ID of the source GOES satellite (G16 or G17).

datetime
Desired datetime in any Ymd H [MS] format or POSIXct

endTime
Desired ending time in any Ymd H [MS] format or POSIXct

timezone
Timezone used to interpret datetime and endTime

isJulian
Logical specifying if datetime (and optionally endTime) are Julian formatted

verbose
Logical parameter specifying whether download status should be shown.

baseUrl
Base URL for data queries.

Value

Invisibly returns a vector of local files matching the requested datetime.

See also

Examples

# \donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Single hour
files <- goesaodc_downloadAOD("G17", "2019-09-06 18")
print(files)

#>  [1] "OR_ABI-L2-AODC-M6_G17_s20192491801196_e20192491803569_c20192491805437.nc"
#>  [2] "OR_ABI-L2-AODC-M6_G17_s20192491806196_e20192491808569_c20192491810456.nc"
#>  [3] "OR_ABI-L2-AODC-M6_G17_s20192491811196_e20192491813569_c20192491815453.nc"
#>  [4] "OR_ABI-L2-AODC-M6_G17_s20192491816196_e20192491818569_c20192491820468.nc"
#>  [5] "OR_ABI-L2-AODC-M6_G17_s20192491821196_e20192491823569_c20192491825491.nc"
#>  [6] "OR_ABI-L2-AODC-M6_G17_s20192491826196_e20192491828569_c20192491830494.nc"
#>  [7] "OR_ABI-L2-AODC-M6_G17_s20192491831196_e20192491833569_c20192491835517.nc"
#>  [8] "OR_ABI-L2-AODC-M6_G17_s20192491836196_e20192491838569_c20192491840517.nc"
#>  [9] "OR_ABI-L2-AODC-M6_G17_s20192491841196_e20192491843569_c20192491845528.nc"
#> [10] "OR_ABI-L2-AODC-M6_G17_s20192491846196_e20192491848569_c20192491850525.nc"
#> [11] "OR_ABI-L2-AODC-M6_G17_s20192491851196_e20192491853569_c20192491855529.nc"
#> [12] "OR_ABI-L2-AODC-M6_G17_s20192491856196_e20192491858569_c20192491900549.nc"

# Specific time range
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "2019-09-06 08:00",
  endTime = "2019-09-06 12:00",
  timezone = "America/Los_Angeles"
)
print(files)

#>  [1] "OR_ABI-L2-AODC-M6_G16_s20192491501093_e20192491503466_c20192491510124.nc"
#>  [2] "OR_ABI-L2-AODC-M6_G16_s20192491506093_e20192491508466_c20192491515087.nc"
#>  [3] "OR_ABI-L2-AODC-M6_G16_s20192491511094_e20192491513466_c20192491520123.nc"
#>  [4] "OR_ABI-L2-AODC-M6_G16_s20192491516094_e20192491518466_c20192491525124.nc"
#>  [5] "OR_ABI-L2-AODC-M6_G16_s20192491521094_e20192491523467_c20192491530105.nc"
#>  [6] "OR_ABI-L2-AODC-M6_G16_s20192491526094_e20192491528467_c20192491535106.nc"
#>  [7] "OR_ABI-L2-AODC-M6_G16_s20192491531094_e20192491533467_c20192491540114.nc"
#>  [8] "OR_ABI-L2-AODC-M6_G16_s20192491536094_e20192491538467_c20192491545088.nc"
#>  [9] "OR_ABI-L2-AODC-M6_G16_s20192491541094_e20192491543467_c20192491550116.nc"
#> [10] "OR_ABI-L2-AODC-M6_G16_s20192491546094_e20192491548467_c20192491555121.nc"
#> [11] "OR_ABI-L2-AODC-M6_G16_s20192491551094_e20192491553467_c20192491600096.nc"
#> [12] "OR_ABI-L2-AODC-M6_G16_s20192491556094_e20192491558467_c20192491605110.nc"
#> [13] "OR_ABI-L2-AODC-M6_G16_s20192491601094_e20192491603467_c20192491610121.nc"
#> [14] "OR_ABI-L2-AODC-M6_G16_s20192491606094_e20192491608467_c20192491615088.nc"
#> [15] "OR_ABI-L2-AODC-M6_G16_s20192491611094_e20192491613467_c20192491620114.nc"
#> [16] "OR_ABI-L2-AODC-M6_G16_s20192491616094_e20192491618467_c20192491625131.nc"
#> [17] "OR_ABI-L2-AODC-M6_G16_s20192491621094_e20192491623467_c20192491630089.nc"
#> [18] "OR_ABI-L2-AODC-M6_G16_s20192491626094_e20192491628467_c20192491635120.nc"
#> [19] "OR_ABI-L2-AODC-M6_G16_s20192491631094_e20192491633467_c20192491640112.nc"
#> [20] "OR_ABI-L2-AODC-M6_G16_s20192491636094_e20192491638467_c20192491645095.nc"
#> [21] "OR_ABI-L2-AODC-M6_G16_s20192491641094_e20192491643467_c20192491650101.nc"
#> [22] "OR_ABI-L2-AODC-M6_G16_s20192491646094_e20192491648467_c20192491655131.nc"
#> [23] "OR_ABI-L2-AODC-M6_G16_s20192491651094_e20192491653467_c20192491700096.nc"
#> [24] "OR_ABI-L2-AODC-M6_G16_s20192491656094_e20192491658467_c20192491705115.nc"
#> [25] "OR_ABI-L2-AODC-M6_G16_s20192491701094_e20192491703467_c20192491710110.nc"
#> [26] "OR_ABI-L2-AODC-M6_G16_s20192491706094_e20192491708467_c20192491715102.nc"
#> [27] "OR_ABI-L2-AODC-M6_G16_s20192491711095_e20192491713467_c20192491720125.nc"
#> [28] "OR_ABI-L2-AODC-M6_G16_s20192491716095_e20192491718467_c20192491725125.nc"
#> [29] "OR_ABI-L2-AODC-M6_G16_s20192491721095_e20192491723468_c20192491730105.nc"
#> [30] "OR_ABI-L2-AODC-M6_G16_s20192491726095_e20192491728468_c20192491735127.nc"
#> [31] "OR_ABI-L2-AODC-M6_G16_s20192491731095_e20192491733468_c20192491740122.nc"
#> [32] "OR_ABI-L2-AODC-M6_G16_s20192491736095_e20192491738468_c20192491745098.nc"
#> [33] "OR_ABI-L2-AODC-M6_G16_s20192491741095_e20192491743468_c20192491750108.nc"
#> [34] "OR_ABI-L2-AODC-M6_G16_s20192491746095_e20192491748468_c20192491755131.nc"
#> [35] "OR_ABI-L2-AODC-M6_G16_s20192491751095_e20192491753468_c20192491800111.nc"
#> [36] "OR_ABI-L2-AODC-M6_G16_s20192491756095_e20192491758468_c20192491805104.nc"
#> [37] "OR_ABI-L2-AODC-M6_G16_s20192491801095_e20192491803468_c20192491810129.nc"
#> [38] "OR_ABI-L2-AODC-M6_G16_s20192491806095_e20192491808468_c20192491815101.nc"
#> [39] "OR_ABI-L2-AODC-M6_G16_s20192491811095_e20192491813468_c20192491820128.nc"
#> [40] "OR_ABI-L2-AODC-M6_G16_s20192491816095_e20192491818468_c20192491825122.nc"
#> [41] "OR_ABI-L2-AODC-M6_G16_s20192491821095_e20192491823468_c20192491830111.nc"
#> [42] "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc"
#> [43] "OR_ABI-L2-AODC-M6_G16_s20192491831095_e20192491833468_c20192491840122.nc"
#> [44] "OR_ABI-L2-AODC-M6_G16_s20192491836095_e20192491838468_c20192491845095.nc"
#> [45] "OR_ABI-L2-AODC-M6_G16_s20192491841095_e20192491843468_c20192491850120.nc"
#> [46] "OR_ABI-L2-AODC-M6_G16_s20192491846095_e20192491848468_c20192491855133.nc"
#> [47] "OR_ABI-L2-AODC-M6_G16_s20192491851095_e20192491853468_c20192491900103.nc"
#> [48] "OR_ABI-L2-AODC-M6_G16_s20192491856095_e20192491858468_c20192491905110.nc"
#> [49] "OR_ABI-L2-AODC-M6_G16_s20192491901095_e20192491903468_c20192491910119.nc"
#> [50] "OR_ABI-L2-AODC-M6_G16_s20192491906095_e20192491908468_c20192491915112.nc"
#> [51] "OR_ABI-L2-AODC-M6_G16_s20192491911096_e20192491913468_c20192491919214.nc"
#> [52] "OR_ABI-L2-AODC-M6_G16_s20192491916096_e20192491918468_c20192491923198.nc"
#> [53] "OR_ABI-L2-AODC-M6_G16_s20192491921096_e20192491923469_c20192491927063.nc"
#> [54] "OR_ABI-L2-AODC-M6_G16_s20192491926096_e20192491928469_c20192491931399.nc"
#> [55] "OR_ABI-L2-AODC-M6_G16_s20192491931096_e20192491933469_c20192491936389.nc"
#> [56] "OR_ABI-L2-AODC-M6_G16_s20192491936096_e20192491938469_c20192491941380.nc"
#> [57] "OR_ABI-L2-AODC-M6_G16_s20192491941096_e20192491943469_c20192491946332.nc"
#> [58] "OR_ABI-L2-AODC-M6_G16_s20192491946096_e20192491948469_c20192491951448.nc"
#> [59] "OR_ABI-L2-AODC-M6_G16_s20192491951096_e20192491953469_c20192491956325.nc"
#> [60] "OR_ABI-L2-AODC-M6_G16_s20192491956096_e20192491958469_c20192492001307.nc"

# Julian timestamp
files <- goesaodc_downloadAOD(
  satID = "G16",
  datetime = "201924915",
  isJulian = TRUE
)
print(files)

#>  [1] "OR_ABI-L2-AODC-M6_G16_s20192491501093_e20192491503466_c20192491510124.nc"
#>  [2] "OR_ABI-L2-AODC-M6_G16_s20192491506093_e20192491508466_c20192491515087.nc"
#>  [3] "OR_ABI-L2-AODC-M6_G16_s20192491511094_e20192491513466_c20192491520123.nc"
#>  [4] "OR_ABI-L2-AODC-M6_G16_s20192491516094_e20192491518466_c20192491525124.nc"
#>  [5] "OR_ABI-L2-AODC-M6_G16_s20192491521094_e20192491523467_c20192491530105.nc"
#>  [6] "OR_ABI-L2-AODC-M6_G16_s20192491526094_e20192491528467_c20192491535106.nc"
#>  [7] "OR_ABI-L2-AODC-M6_G16_s20192491531094_e20192491533467_c20192491540114.nc"
#>  [8] "OR_ABI-L2-AODC-M6_G16_s20192491536094_e20192491538467_c20192491545088.nc"
#>  [9] "OR_ABI-L2-AODC-M6_G16_s20192491541094_e20192491543467_c20192491550116.nc"
#> [10] "OR_ABI-L2-AODC-M6_G16_s20192491546094_e20192491548467_c20192491555121.nc"
#> [11] "OR_ABI-L2-AODC-M6_G16_s20192491551094_e20192491553467_c20192491600096.nc"
#> [12] "OR_ABI-L2-AODC-M6_G16_s20192491556094_e20192491558467_c20192491605110.nc"

# }