Download GOES-16 or GOES-17 AOD data

Download all GOES 16 or 17 NetCDF files for the given datetime to the directory specified by setSatelliteDataDir(). If datetime is specified to the hour and fullDay is not explicitly set to TRUE', only files for that hour will be downloaded. If the optional endTime is specified, all files that exist for the time period between datetime and endTime will be downloaded. If timezone is not specified, "UTC" timezone is assumed.

goesaodc_downloadAOD(satID = NULL, datetime = NULL, endTime = NULL,
  timezone = "UTC", julian = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/")

Arguments

satID
ID of the source GOES satellite (G16 or G17).

datetime
Desired datetime in any Ymd H [MS] format or POSIXct

endTime
Desired ending time in any Ymd H [MS] format or POSIXct

timezone
Timezone used to interpret datetime and endTime

julian
Logical specifying if datetime (and optionally endTime) are Julian formatted

baseUrl
Base URL for data queries.

Value

Vector of downloaded filepaths.

See also

Examples

# \donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

goesaodc_downloadAOD(satID = "G17",
datetime = "2019-09-06 18")

goesaodc_downloadAOD(satID = "G16",
datetime = "2019-09-06 08:00",
endTime = "2019-09-06 12:00",
timezone = "America/Los_Angeles")

goesaodc_downloadAOD(satID = "G16",
datetime = "201924915",
julian = TRUE)

# }