List available GOES AOD files for a specified date and hour

Retrieve a list of GOES AOD files available in the satelliteDataDir or at baseUrl for a specified date and hour.

goesaodc_listFiles(
  satID = NULL,
  datetime = NULL,
  endTime = NULL,
  useRemote = FALSE,
  timezone = "UTC",
  isJulian = FALSE,
  baseUrl = "https://tools-1.airfire.org/Satellite/"
)

Arguments

satID
ID of the source GOES satellite (G16 or G17).

datetime
Desired datetime in any Ymd H [MS] format or POSIXct

endTime
Desired ending time in any Ymd H [MS] format or POSIXct

useRemote
Logical specifying whether to look for files in getSatelliteDataDir() or baseUrl.

timezone
Timezone used to interpret datetime and endTime

isJulian
Logical specifying if datetime (and optionally endTime) are Julian formatted

baseUrl
Base URL for data queries.

Value

Vector of filenames.

Note

All files for a particular hour will be returned even if the incoming startdate or jdate is specified to the minute or second. By default, the local directory set by setSatelliteDir() is searched. useRemote = TRUE will search baseURL for all files that are available within the specified timeframe.

Examples

# \donttest{
library(MazamaSatelliteUtils)
setSatelliteDataDir("~/Data/Satellite")

# Local files for a UTC hour
goesaodc_listFiles("G17", "2019-09-06 18")

#>  [1] "OR_ABI-L2-AODC-M6_G17_s20192491801196_e20192491803569_c20192491805437.nc"
#>  [2] "OR_ABI-L2-AODC-M6_G17_s20192491806196_e20192491808569_c20192491810456.nc"
#>  [3] "OR_ABI-L2-AODC-M6_G17_s20192491811196_e20192491813569_c20192491815453.nc"
#>  [4] "OR_ABI-L2-AODC-M6_G17_s20192491816196_e20192491818569_c20192491820468.nc"
#>  [5] "OR_ABI-L2-AODC-M6_G17_s20192491821196_e20192491823569_c20192491825491.nc"
#>  [6] "OR_ABI-L2-AODC-M6_G17_s20192491826196_e20192491828569_c20192491830494.nc"
#>  [7] "OR_ABI-L2-AODC-M6_G17_s20192491831196_e20192491833569_c20192491835517.nc"
#>  [8] "OR_ABI-L2-AODC-M6_G17_s20192491836196_e20192491838569_c20192491840517.nc"
#>  [9] "OR_ABI-L2-AODC-M6_G17_s20192491841196_e20192491843569_c20192491845528.nc"
#> [10] "OR_ABI-L2-AODC-M6_G17_s20192491846196_e20192491848569_c20192491850525.nc"
#> [11] "OR_ABI-L2-AODC-M6_G17_s20192491851196_e20192491853569_c20192491855529.nc"
#> [12] "OR_ABI-L2-AODC-M6_G17_s20192491856196_e20192491858569_c20192491900549.nc"

# Remote files for a time range in local time
goesaodc_listFiles(
  satID = "G16",
  datetime = "2019-09-06 06:00",
  endTime = "2019-09-06 18:00",
  timezone = "America/Los_Angeles",
  useRemote = TRUE
)

#>   [1] "OR_ABI-L2-AODC-M6_G16_s20192491301093_e20192491303466_c20192491310120.nc"
#>   [2] "OR_ABI-L2-AODC-M6_G16_s20192491306093_e20192491308466_c20192491315092.nc"
#>   [3] "OR_ABI-L2-AODC-M6_G16_s20192491311093_e20192491313466_c20192491320109.nc"
#>   [4] "OR_ABI-L2-AODC-M6_G16_s20192491316093_e20192491318466_c20192491325115.nc"
#>   [5] "OR_ABI-L2-AODC-M6_G16_s20192491321093_e20192491323466_c20192491330118.nc"
#>   [6] "OR_ABI-L2-AODC-M6_G16_s20192491326093_e20192491328466_c20192491335103.nc"
#>   [7] "OR_ABI-L2-AODC-M6_G16_s20192491331093_e20192491333466_c20192491340118.nc"
#>   [8] "OR_ABI-L2-AODC-M6_G16_s20192491336093_e20192491338466_c20192491345081.nc"
#>   [9] "OR_ABI-L2-AODC-M6_G16_s20192491341093_e20192491343466_c20192491350104.nc"
#>  [10] "OR_ABI-L2-AODC-M6_G16_s20192491346093_e20192491348466_c20192491355108.nc"
#>  [11] "OR_ABI-L2-AODC-M6_G16_s20192491351093_e20192491353466_c20192491400106.nc"
#>  [12] "OR_ABI-L2-AODC-M6_G16_s20192491356093_e20192491358466_c20192491405105.nc"
#>  [13] "OR_ABI-L2-AODC-M6_G16_s20192491401093_e20192491403466_c20192491410135.nc"
#>  [14] "OR_ABI-L2-AODC-M6_G16_s20192491406093_e20192491408466_c20192491415092.nc"
#>  [15] "OR_ABI-L2-AODC-M6_G16_s20192491411093_e20192491413466_c20192491420120.nc"
#>  [16] "OR_ABI-L2-AODC-M6_G16_s20192491416093_e20192491418466_c20192491425128.nc"
#>  [17] "OR_ABI-L2-AODC-M6_G16_s20192491421093_e20192491423466_c20192491430114.nc"
#>  [18] "OR_ABI-L2-AODC-M6_G16_s20192491426093_e20192491428466_c20192491435109.nc"
#>  [19] "OR_ABI-L2-AODC-M6_G16_s20192491431093_e20192491433466_c20192491440123.nc"
#>  [20] "OR_ABI-L2-AODC-M6_G16_s20192491436093_e20192491438466_c20192491445090.nc"
#>  [21] "OR_ABI-L2-AODC-M6_G16_s20192491441093_e20192491443466_c20192491450106.nc"
#>  [22] "OR_ABI-L2-AODC-M6_G16_s20192491446093_e20192491448466_c20192491455129.nc"
#>  [23] "OR_ABI-L2-AODC-M6_G16_s20192491451093_e20192491453466_c20192491500114.nc"
#>  [24] "OR_ABI-L2-AODC-M6_G16_s20192491456093_e20192491458466_c20192491505098.nc"
#>  [25] "OR_ABI-L2-AODC-M6_G16_s20192491501093_e20192491503466_c20192491510124.nc"
#>  [26] "OR_ABI-L2-AODC-M6_G16_s20192491506093_e20192491508466_c20192491515087.nc"
#>  [27] "OR_ABI-L2-AODC-M6_G16_s20192491511094_e20192491513466_c20192491520123.nc"
#>  [28] "OR_ABI-L2-AODC-M6_G16_s20192491516094_e20192491518466_c20192491525124.nc"
#>  [29] "OR_ABI-L2-AODC-M6_G16_s20192491521094_e20192491523467_c20192491530105.nc"
#>  [30] "OR_ABI-L2-AODC-M6_G16_s20192491526094_e20192491528467_c20192491535106.nc"
#>  [31] "OR_ABI-L2-AODC-M6_G16_s20192491531094_e20192491533467_c20192491540114.nc"
#>  [32] "OR_ABI-L2-AODC-M6_G16_s20192491536094_e20192491538467_c20192491545088.nc"
#>  [33] "OR_ABI-L2-AODC-M6_G16_s20192491541094_e20192491543467_c20192491550116.nc"
#>  [34] "OR_ABI-L2-AODC-M6_G16_s20192491546094_e20192491548467_c20192491555121.nc"
#>  [35] "OR_ABI-L2-AODC-M6_G16_s20192491551094_e20192491553467_c20192491600096.nc"
#>  [36] "OR_ABI-L2-AODC-M6_G16_s20192491556094_e20192491558467_c20192491605110.nc"
#>  [37] "OR_ABI-L2-AODC-M6_G16_s20192491601094_e20192491603467_c20192491610121.nc"
#>  [38] "OR_ABI-L2-AODC-M6_G16_s20192491606094_e20192491608467_c20192491615088.nc"
#>  [39] "OR_ABI-L2-AODC-M6_G16_s20192491611094_e20192491613467_c20192491620114.nc"
#>  [40] "OR_ABI-L2-AODC-M6_G16_s20192491616094_e20192491618467_c20192491625131.nc"
#>  [41] "OR_ABI-L2-AODC-M6_G16_s20192491621094_e20192491623467_c20192491630089.nc"
#>  [42] "OR_ABI-L2-AODC-M6_G16_s20192491626094_e20192491628467_c20192491635120.nc"
#>  [43] "OR_ABI-L2-AODC-M6_G16_s20192491631094_e20192491633467_c20192491640112.nc"
#>  [44] "OR_ABI-L2-AODC-M6_G16_s20192491636094_e20192491638467_c20192491645095.nc"
#>  [45] "OR_ABI-L2-AODC-M6_G16_s20192491641094_e20192491643467_c20192491650101.nc"
#>  [46] "OR_ABI-L2-AODC-M6_G16_s20192491646094_e20192491648467_c20192491655131.nc"
#>  [47] "OR_ABI-L2-AODC-M6_G16_s20192491651094_e20192491653467_c20192491700096.nc"
#>  [48] "OR_ABI-L2-AODC-M6_G16_s20192491656094_e20192491658467_c20192491705115.nc"
#>  [49] "OR_ABI-L2-AODC-M6_G16_s20192491701094_e20192491703467_c20192491710110.nc"
#>  [50] "OR_ABI-L2-AODC-M6_G16_s20192491706094_e20192491708467_c20192491715102.nc"
#>  [51] "OR_ABI-L2-AODC-M6_G16_s20192491711095_e20192491713467_c20192491720125.nc"
#>  [52] "OR_ABI-L2-AODC-M6_G16_s20192491716095_e20192491718467_c20192491725125.nc"
#>  [53] "OR_ABI-L2-AODC-M6_G16_s20192491721095_e20192491723468_c20192491730105.nc"
#>  [54] "OR_ABI-L2-AODC-M6_G16_s20192491726095_e20192491728468_c20192491735127.nc"
#>  [55] "OR_ABI-L2-AODC-M6_G16_s20192491731095_e20192491733468_c20192491740122.nc"
#>  [56] "OR_ABI-L2-AODC-M6_G16_s20192491736095_e20192491738468_c20192491745098.nc"
#>  [57] "OR_ABI-L2-AODC-M6_G16_s20192491741095_e20192491743468_c20192491750108.nc"
#>  [58] "OR_ABI-L2-AODC-M6_G16_s20192491746095_e20192491748468_c20192491755131.nc"
#>  [59] "OR_ABI-L2-AODC-M6_G16_s20192491751095_e20192491753468_c20192491800111.nc"
#>  [60] "OR_ABI-L2-AODC-M6_G16_s20192491756095_e20192491758468_c20192491805104.nc"
#>  [61] "OR_ABI-L2-AODC-M6_G16_s20192491801095_e20192491803468_c20192491810129.nc"
#>  [62] "OR_ABI-L2-AODC-M6_G16_s20192491806095_e20192491808468_c20192491815101.nc"
#>  [63] "OR_ABI-L2-AODC-M6_G16_s20192491811095_e20192491813468_c20192491820128.nc"
#>  [64] "OR_ABI-L2-AODC-M6_G16_s20192491816095_e20192491818468_c20192491825122.nc"
#>  [65] "OR_ABI-L2-AODC-M6_G16_s20192491821095_e20192491823468_c20192491830111.nc"
#>  [66] "OR_ABI-L2-AODC-M6_G16_s20192491826095_e20192491828468_c20192491835127.nc"
#>  [67] "OR_ABI-L2-AODC-M6_G16_s20192491831095_e20192491833468_c20192491840122.nc"
#>  [68] "OR_ABI-L2-AODC-M6_G16_s20192491836095_e20192491838468_c20192491845095.nc"
#>  [69] "OR_ABI-L2-AODC-M6_G16_s20192491841095_e20192491843468_c20192491850120.nc"
#>  [70] "OR_ABI-L2-AODC-M6_G16_s20192491846095_e20192491848468_c20192491855133.nc"
#>  [71] "OR_ABI-L2-AODC-M6_G16_s20192491851095_e20192491853468_c20192491900103.nc"
#>  [72] "OR_ABI-L2-AODC-M6_G16_s20192491856095_e20192491858468_c20192491905110.nc"
#>  [73] "OR_ABI-L2-AODC-M6_G16_s20192491901095_e20192491903468_c20192491910119.nc"
#>  [74] "OR_ABI-L2-AODC-M6_G16_s20192491906095_e20192491908468_c20192491915112.nc"
#>  [75] "OR_ABI-L2-AODC-M6_G16_s20192491911096_e20192491913468_c20192491919214.nc"
#>  [76] "OR_ABI-L2-AODC-M6_G16_s20192491916096_e20192491918468_c20192491923198.nc"
#>  [77] "OR_ABI-L2-AODC-M6_G16_s20192491921096_e20192491923469_c20192491927063.nc"
#>  [78] "OR_ABI-L2-AODC-M6_G16_s20192491926096_e20192491928469_c20192491931399.nc"
#>  [79] "OR_ABI-L2-AODC-M6_G16_s20192491931096_e20192491933469_c20192491936389.nc"
#>  [80] "OR_ABI-L2-AODC-M6_G16_s20192491936096_e20192491938469_c20192491941380.nc"
#>  [81] "OR_ABI-L2-AODC-M6_G16_s20192491941096_e20192491943469_c20192491946332.nc"
#>  [82] "OR_ABI-L2-AODC-M6_G16_s20192491946096_e20192491948469_c20192491951448.nc"
#>  [83] "OR_ABI-L2-AODC-M6_G16_s20192491951096_e20192491953469_c20192491956325.nc"
#>  [84] "OR_ABI-L2-AODC-M6_G16_s20192491956096_e20192491958469_c20192492001307.nc"
#>  [85] "OR_ABI-L2-AODC-M6_G16_s20192492001096_e20192492003469_c20192492006360.nc"
#>  [86] "OR_ABI-L2-AODC-M6_G16_s20192492006096_e20192492008469_c20192492011372.nc"
#>  [87] "OR_ABI-L2-AODC-M6_G16_s20192492011096_e20192492013469_c20192492016332.nc"
#>  [88] "OR_ABI-L2-AODC-M6_G16_s20192492016096_e20192492018469_c20192492021445.nc"
#>  [89] "OR_ABI-L2-AODC-M6_G16_s20192492021096_e20192492023469_c20192492026317.nc"
#>  [90] "OR_ABI-L2-AODC-M6_G16_s20192492026096_e20192492028469_c20192492031299.nc"
#>  [91] "OR_ABI-L2-AODC-M6_G16_s20192492031096_e20192492033469_c20192492036320.nc"
#>  [92] "OR_ABI-L2-AODC-M6_G16_s20192492036096_e20192492038469_c20192492041335.nc"
#>  [93] "OR_ABI-L2-AODC-M6_G16_s20192492041096_e20192492043469_c20192492046270.nc"
#>  [94] "OR_ABI-L2-AODC-M6_G16_s20192492046096_e20192492048469_c20192492051349.nc"
#>  [95] "OR_ABI-L2-AODC-M6_G16_s20192492051096_e20192492053469_c20192492056309.nc"
#>  [96] "OR_ABI-L2-AODC-M6_G16_s20192492056096_e20192492058469_c20192492101269.nc"
#>  [97] "OR_ABI-L2-AODC-M6_G16_s20192492101096_e20192492103469_c20192492106337.nc"
#>  [98] "OR_ABI-L2-AODC-M6_G16_s20192492106096_e20192492108469_c20192492111398.nc"
#>  [99] "OR_ABI-L2-AODC-M6_G16_s20192492111096_e20192492113469_c20192492116300.nc"
#> [100] "OR_ABI-L2-AODC-M6_G16_s20192492116096_e20192492118469_c20192492121346.nc"
#> [101] "OR_ABI-L2-AODC-M6_G16_s20192492121097_e20192492123469_c20192492126248.nc"
#> [102] "OR_ABI-L2-AODC-M6_G16_s20192492126097_e20192492128469_c20192492131198.nc"
#> [103] "OR_ABI-L2-AODC-M6_G16_s20192492131097_e20192492133469_c20192492136246.nc"
#> [104] "OR_ABI-L2-AODC-M6_G16_s20192492136097_e20192492138470_c20192492141214.nc"
#> [105] "OR_ABI-L2-AODC-M6_G16_s20192492141097_e20192492143470_c20192492146200.nc"
#> [106] "OR_ABI-L2-AODC-M6_G16_s20192492146097_e20192492148470_c20192492151204.nc"
#> [107] "OR_ABI-L2-AODC-M6_G16_s20192492151097_e20192492153470_c20192492156167.nc"
#> [108] "OR_ABI-L2-AODC-M6_G16_s20192492156097_e20192492158470_c20192492201139.nc"
#> [109] "OR_ABI-L2-AODC-M6_G16_s20192492201097_e20192492203470_c20192492206207.nc"
#> [110] "OR_ABI-L2-AODC-M6_G16_s20192492206097_e20192492208470_c20192492211149.nc"
#> [111] "OR_ABI-L2-AODC-M6_G16_s20192492211097_e20192492213470_c20192492216112.nc"
#> [112] "OR_ABI-L2-AODC-M6_G16_s20192492216097_e20192492218470_c20192492221321.nc"
#> [113] "OR_ABI-L2-AODC-M6_G16_s20192492221097_e20192492223470_c20192492226091.nc"
#> [114] "OR_ABI-L2-AODC-M6_G16_s20192492226097_e20192492228470_c20192492231115.nc"
#> [115] "OR_ABI-L2-AODC-M6_G16_s20192492231097_e20192492233470_c20192492236133.nc"
#> [116] "OR_ABI-L2-AODC-M6_G16_s20192492236097_e20192492238470_c20192492241231.nc"
#> [117] "OR_ABI-L2-AODC-M6_G16_s20192492241097_e20192492243470_c20192492246231.nc"
#> [118] "OR_ABI-L2-AODC-M6_G16_s20192492246097_e20192492248470_c20192492251356.nc"
#> [119] "OR_ABI-L2-AODC-M6_G16_s20192492251097_e20192492253470_c20192492256396.nc"
#> [120] "OR_ABI-L2-AODC-M6_G16_s20192492256097_e20192492258470_c20192492301590.nc"
#> [121] "OR_ABI-L2-AODC-M6_G16_s20192492301097_e20192492303470_c20192492306531.nc"
#> [122] "OR_ABI-L2-AODC-M6_G16_s20192492306097_e20192492308470_c20192492311210.nc"
#> [123] "OR_ABI-L2-AODC-M6_G16_s20192492311097_e20192492313470_c20192492316185.nc"
#> [124] "OR_ABI-L2-AODC-M6_G16_s20192492316097_e20192492318470_c20192492321436.nc"
#> [125] "OR_ABI-L2-AODC-M6_G16_s20192492321097_e20192492323470_c20192492326333.nc"
#> [126] "OR_ABI-L2-AODC-M6_G16_s20192492326097_e20192492328470_c20192492331254.nc"
#> [127] "OR_ABI-L2-AODC-M6_G16_s20192492331098_e20192492333470_c20192492335582.nc"
#> [128] "OR_ABI-L2-AODC-M6_G16_s20192492336098_e20192492338470_c20192492341053.nc"
#> [129] "OR_ABI-L2-AODC-M6_G16_s20192492341098_e20192492343470_c20192492346088.nc"
#> [130] "OR_ABI-L2-AODC-M6_G16_s20192492346098_e20192492348470_c20192492351376.nc"
#> [131] "OR_ABI-L2-AODC-M6_G16_s20192492351098_e20192492353471_c20192492356190.nc"
#> [132] "OR_ABI-L2-AODC-M6_G16_s20192492356098_e20192492358471_c20192500001041.nc"
#> [133] "OR_ABI-L2-AODC-M6_G16_s20192500001098_e20192500003471_c20192500006231.nc"
#> [134] "OR_ABI-L2-AODC-M6_G16_s20192500006098_e20192500008471_c20192500011414.nc"
#> [135] "OR_ABI-L2-AODC-M6_G16_s20192500011098_e20192500013471_c20192500016335.nc"
#> [136] "OR_ABI-L2-AODC-M6_G16_s20192500016098_e20192500018471_c20192500021182.nc"
#> [137] "OR_ABI-L2-AODC-M6_G16_s20192500021098_e20192500023471_c20192500025421.nc"
#> [138] "OR_ABI-L2-AODC-M6_G16_s20192500026098_e20192500028471_c20192500030401.nc"
#> [139] "OR_ABI-L2-AODC-M6_G16_s20192500031098_e20192500033471_c20192500035519.nc"
#> [140] "OR_ABI-L2-AODC-M6_G16_s20192500036098_e20192500038471_c20192500040453.nc"
#> [141] "OR_ABI-L2-AODC-M6_G16_s20192500041098_e20192500043471_c20192500045047.nc"
#> [142] "OR_ABI-L2-AODC-M6_G16_s20192500046098_e20192500048471_c20192500049450.nc"
#> [143] "OR_ABI-L2-AODC-M6_G16_s20192500051098_e20192500053471_c20192500054591.nc"
#> [144] "OR_ABI-L2-AODC-M6_G16_s20192500056098_e20192500058471_c20192500100083.nc"
#> [145] "OR_ABI-L2-AODC-M6_G16_s20192500101098_e20192500103471_c20192500105112.nc"
#> [146] "OR_ABI-L2-AODC-M6_G16_s20192500106098_e20192500108471_c20192500110287.nc"
#> [147] "OR_ABI-L2-AODC-M6_G16_s20192500111098_e20192500113471_c20192500115330.nc"
#> [148] "OR_ABI-L2-AODC-M6_G16_s20192500116098_e20192500118471_c20192500120334.nc"
#> [149] "OR_ABI-L2-AODC-M6_G16_s20192500121098_e20192500123471_c20192500125387.nc"
#> [150] "OR_ABI-L2-AODC-M6_G16_s20192500126098_e20192500128471_c20192500130365.nc"
#> [151] "OR_ABI-L2-AODC-M6_G16_s20192500131098_e20192500133471_c20192500135220.nc"
#> [152] "OR_ABI-L2-AODC-M6_G16_s20192500136098_e20192500138471_c20192500139504.nc"
#> [153] "OR_ABI-L2-AODC-M6_G16_s20192500141098_e20192500143471_c20192500144411.nc"
#> [154] "OR_ABI-L2-AODC-M6_G16_s20192500146098_e20192500148471_c20192500149409.nc"
#> [155] "OR_ABI-L2-AODC-M6_G16_s20192500151099_e20192500153471_c20192500154417.nc"
#> [156] "OR_ABI-L2-AODC-M6_G16_s20192500156099_e20192500158471_c20192500159373.nc"

# }